requirements.txt
openai>=1.12.0,<2.0.0
mistune>=3.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0 